+++ include/Events/ProjectileExplosionEventData.hpp
@@ -7,7 +7,7 @@ using std::string;
 
 class ProjectileExplosionEventData : public EventDataInterface{
 public:
-  string projectileID;
+  long long projectileID;
   float timeStamp;
 
   ProjectileExplosionEventData(long long  projectileID, float timeStamp);