+++ src/AllyCompView.cpp
@@ -8,7 +8,28 @@ AllyCompView::AllyCompView(shared_ptr<EventManager> eventManager, shared_ptr<Tex
   this -> playingScreenHeader = playingScreenHeader;
 }
 
-void AllyCompView::updateCompView(float deltaS){
+void AllyCompView::updateAllyCompView(float deltaS){
   if(gameLogic->getGameState()==State::Playing){
+    updateProjectiles(deltaS);
   }
+}
+/*
+ * Call the update methods for all the projectiles which will move them closer to their target
+ * if they hit
+ */
+void AllyCompView::updateProjectiles(float deltaS){
+  unordered_map<long long, shared_ptr<ActorInterface>> allProjectiles = gameLogic -> getFiredProjectiles();
+
+  for(auto it : allProjectiles){
+
+    shared_ptr<ActorInterface> projectile = it.second;
+
+    Projectile* convertedProjeectile = dynamic_cast<Projectile*>(projectile.get());
+
+    convertedProjeectile -> update(deltaS);
+  }
+}
+
+void AllyCompView::moveAllyUnits(float deltaS){
+
 }