+++ src/Projectiles/GravyProjectile.cpp
@@ -8,5 +8,11 @@ GravyProjectile::GravyProjectile(shared_ptr<TextLoader> textLoader, shared_ptr<E
   this -> radius = textLoader->getInteger(string("IDS_G_AR"));;
   this -> actorTypeID = textLoader->getTypeID(string("IDS_GP"));
   this -> textures = textureLoader -> getTexture(actorTypeID);
-  this -> initSprite();
+  //set the initial sprite texture
+  this ->current_sprite = 0;
+  //load in the initial texture for sizing
+  initSprite();
+
+  //set the sprite for the actor to have a position that is equivalent to its center
+  setToCenter();
 }