+++ src/Projectiles/MeatLoversProjectile.cpp
@@ -8,5 +8,11 @@ MeatLoversProjectile::MeatLoversProjectile(shared_ptr<TextLoader> textLoader, sh
   this -> radius = textLoader->getInteger(string("IDS_ML_AR"));;
   this -> actorTypeID = textLoader->getTypeID(string("IDS_MLP"));
   this -> textures = textureLoader -> getTexture(actorTypeID);
-  this -> initSprite();
+  //set the initial sprite texture
+  this ->current_sprite = 0;
+  //load in the initial texture for sizing
+  initSprite();
+
+  //set the sprite for the actor to have a position that is equivalent to its center
+  setToCenter();
 }