+++ src/Projectiles/SodaProjectile.cpp
@@ -9,5 +9,11 @@ SodaProjectile::SodaProjectile(shared_ptr<TextLoader> textLoader, shared_ptr<Eve
   this -> radius = textLoader->getInteger(string("IDS_SO_AR"));;
   this -> actorTypeID = textLoader->getTypeID(string("IDS_SOP"));
   this -> textures = textureLoader -> getTexture(actorTypeID);
-  this -> initSprite();
+  //set the initial sprite texture
+  this ->current_sprite = 0;
+  //load in the initial texture for sizing
+  initSprite();
+
+  //set the sprite for the actor to have a position that is equivalent to its center
+  setToCenter();
 }