+++ src/Units/AverageKidUnit.cpp
@@ -14,5 +14,11 @@ AverageKidUnit::AverageKidUnit(shared_ptr<TextLoader> textLoader, shared_ptr<Eve
   this->lunchMoney=textLoader->getInteger(string("IDS_AK_LM"));;
   this->actorTypeID=textLoader->getTypeID(string("IDS_AKU"));
   this -> textures = textureLoader -> getTexture(actorTypeID);
-  this -> initSprite();
+  //set the initial sprite texture
+  this ->current_sprite = 0;
+  //load in the initial texture for sizing
+  initSprite();
+
+  //set the sprite for the actor to have a position that is equivalent to its center
+  setToCenter();
 }